Added Basic Error Checking
This update checks to see if the person has pressed the operation buttons too many times. If the user tries to divide by zero. If the user tries to perform an operation before pressing a number. And if the user tries to subtract a larger number from a smaller number or divide a smaller number by a bigger number. The check for pressing the equals button too early is still a malfunctioning...